add progress saving during all2all
Created a new function all2all3D_saveIntermediate that will save
progress after each row is completed. This is slower and creates
an unncessary file, but will allow for the slowest (sometimes
days) step of the procedure to be interupted.  If interupted, the
function will load the saved intermediate file and start from the
first non-empty row.